Split actions.panel out of actions.ui.
