+++ src/js/actions/ui.js
@@ -27,50 +27,17 @@ define(function (require, exports) {
     var Promise = require("bluebird"),
         _ = require("lodash");
 
-    var adapter = require("adapter"),
-        descriptor = require("adapter").ps.descriptor,
+    var descriptor = require("adapter").ps.descriptor,
         documentLib = require("adapter").lib.document,
         adapterUI = require("adapter").ps.ui,
-        adapterOS = require("adapter").os,
-        appLib = require("adapter").lib.application;
+        adapterOS = require("adapter").os;
 
     var Bounds = require("js/models/bounds"),
         events = require("js/events"),
         locks = require("js/locks"),
         shortcuts = require("./shortcuts"),
-        preferences = require("./preferences"),
         synchronization = require("js/util/synchronization"),
-        system = require("js/util/system"),
-        headlights = require("js/util/headlights"),
-        uiUtil = require("js/util/ui");
-
-    /**
-     * Tooltip property key that determines the delay until tooltips are shown.
-     *
-     * @const
-     * @private
-     * @type {number}
-     */
-    var TOOLTIP_TIME_KEY = "ui.tooltip.delay.coldToHot";
-
-    /**
-     * The default value for the tooltip coldToHot delay
-     *
-     * @const
-     * @private
-     * @type {number}
-     */
-    var DEFAULT_TOOLTIP_TIME = 0.75;
-
-    /**
-     * A "sufficiently" large value tooltip coldToHot value which effectively
-     * disables tooltips.
-     *
-     * @const
-     * @private
-     * @type {number}
-     */
-    var DISABLED_TOOLTIP_TIME = 9999;
+        system = require("js/util/system");
 
     /**
      * List of zoom increments to fit into
@@ -85,24 +52,6 @@ define(function (require, exports) {
     ];
 
     /**
-     * Key for accessing the reference point from preferences storage.
-     *
-     * @const
-     * @private
-     * @type {string}
-     */
-    var REFERENCE_POINT_PREFS_KEY = "referencePoint";
-
-    /**
-     * The default reference point set
-     *
-     * @const
-     * @private
-     * @type {string}
-     */
-    var DEFAULT_REFERENCE_POINT = "lt";
-
-    /**
      * Document properties needed to update the window transform
      *
      * @private
@@ -113,72 +62,6 @@ define(function (require, exports) {
         "viewTransform",
         "zoom"
     ];
-
-    /**
-     * Globally enable tooltips.
-     *
-     * @return {Promise}
-     */
-    var enableTooltips = function () {
-        return adapter.setPropertyValue(TOOLTIP_TIME_KEY, DEFAULT_TOOLTIP_TIME);
-    };
-    enableTooltips.action = {
-        writes: [locks.PS_APP]
-    };
-
-    /**
-     * Globally disable tooltips and clear any current tooltip.
-     *
-     * @return {Promise}
-     */
-    var disableTooltips = function () {
-        return adapter.setPropertyValue(TOOLTIP_TIME_KEY, DISABLED_TOOLTIP_TIME).then(function () {
-            adapterOS.setTooltip("");
-        });
-    };
-    disableTooltips.action = {
-        writes: [locks.PS_APP]
-    };
-
-    /**
-     * Toggle pinned toolbar
-     *
-     * @return {Promise}
-     */
-    var togglePinnedToolbar = function () {
-        var preferenceState = this.flux.store("preferences").getState(),
-            toolbarPinned = preferenceState.get("toolbarPinned", true);
-
-        var newToolbarPinned = !toolbarPinned;
-
-        return this.transfer(preferences.setPreference, "toolbarPinned", newToolbarPinned);
-    };
-    togglePinnedToolbar.action = {
-        reads: [],
-        writes: [locks.JS_PREF],
-        transfers: [preferences.setPreference]
-    };
-
-    /**
-    * Toggle small screen mode
-    *
-    * @return {Promise}
-    */
-    var toggleSingleColumnMode = function () {
-        var preferenceState = this.flux.store("preferences").getState(),
-            singleColumnModeEnabled = preferenceState.get("singleColumnModeEnabled", false);
-
-        var newsingleColumnModeEnabled = !singleColumnModeEnabled;
-
-        headlights.logEvent("user-interface", "panels", "single-column-mode-" + newsingleColumnModeEnabled);
-
-        return this.transfer(preferences.setPreference, "singleColumnModeEnabled", newsingleColumnModeEnabled);
-    };
-    toggleSingleColumnMode.action = {
-        reads: [],
-        writes: [locks.JS_PREF],
-        transfers: [preferences.setPreference]
-    };
 
     /**
      * Query Photoshop for the curent window transform and emit a
@@ -222,41 +105,6 @@ define(function (require, exports) {
     };
 
     /**
-     * Using the center offsets, creates a cloaking rectangle on the canvas outside panels
-     * that will be blitted out during scroll events
-     *
-     * @return {Promise}
-     */
-    var setOverlayCloaking = function () {
-        var panelStore = this.flux.store("panel"),
-            cloakRect = panelStore.getCloakRect();
-
-        return adapterUI.setOverlayCloaking(cloakRect, ["scroll"], "afterPaint");
-    };
-    setOverlayCloaking.action = {
-        reads: [locks.JS_PANEL],
-        writes: [locks.PS_APP],
-        modal: true
-    };
-
-    /**
-     * Cloak the non-UI portion of the screen immediately, redrawing on the
-     * next repaint.
-     *
-     * @return {Promise}
-     */
-    var cloak = function () {
-        var panelStore = this.flux.store("panel"),
-            cloakRect = panelStore.getCloakRect();
-
-        return adapterUI.setOverlayCloaking(cloakRect, "immediate", "afterPaint");
-    };
-    cloak.action = {
-        reads: [locks.JS_PANEL],
-        writes: [locks.PS_APP]
-    };
-
-    /**
      * Directly emit a TRANSFORM_UPDATED event with the given value.
      *
      * @return {Promise}
@@ -281,72 +129,6 @@ define(function (require, exports) {
         reads: [locks.PS_APP],
         writes: [locks.JS_UI],
         modal: true
-    };
-
-    /**
-     * Parse the panel size information and dispatch the PANELS_RESIZED ui event
-     *
-     * @param {{toolbarWidth: number=, panelWidth: number=, headerHeight: number=}} sizes
-     * @return {Promise}
-     */
-    var updatePanelSizes = function (sizes) {
-        var transformPromise = this.transfer(updateTransform),
-            dispatchPromise = this.dispatchAsync(events.panel.PANELS_RESIZED, sizes)
-            .bind(this)
-                .then(function () {
-                    var centerOffsets = this.flux.store("panel").getState().centerOffsets;
-                    return adapterUI.setOverlayOffsets(centerOffsets);
-                })
-                .then(function () {
-                    return this.transfer(setOverlayCloaking);
-                });
-
-        return Promise.join(transformPromise, dispatchPromise);
-    };
-    updatePanelSizes.action = {
-        reads: [],
-        writes: [locks.JS_PANEL, locks.PS_APP],
-        transfers: [setOverlayCloaking, updateTransform],
-        modal: true
-    };
-
-    /**
-     * Set the overlay offsets in PS in anticipation of opening/creating the
-     * first document (i.e., from a state in which there are no open documents).
-     * This is used, e.g., to ensure that the offsets account for the UI columns
-     * that will be shown once the document is open. See #1999 for more details.
-     *
-     * @return {Promise}
-     */
-    var setOverlayOffsetsForFirstDocument = function () {
-        var flux = this.flux,
-            applicationStore = flux.store("application");
-
-        if (applicationStore.getDocumentCount() > 0) {
-            return Promise.resolve();
-        }
-
-        var preferencesStore = flux.store("preferences"),
-            panelStore = flux.store("panel"),
-            preferences = preferencesStore.getState(),
-            columnCount = 0;
-
-        if (preferences.get(panelStore.components.LAYERS_LIBRARY_COL, true)) {
-            columnCount++;
-        }
-
-        if (preferences.get(panelStore.components.PROPERTIES_COL, true)) {
-            columnCount++;
-        }
-
-        var centerOffsets = panelStore.getCenterOffsets(columnCount);
-
-        return adapterUI.setOverlayOffsets(centerOffsets);
-    };
-    setOverlayOffsetsForFirstDocument.action = {
-        reads: [locks.JS_PREF, locks.JS_APP],
-        writes: [locks.PS_APP],
-        transfers: []
     };
 
     /**
@@ -558,69 +340,20 @@ define(function (require, exports) {
     };
 
     /**
-     * Set the global resize reference point.
-     *
-     * @param {string} referencePoint Two character string denoting the active reference point [lmr][tcb]
-     * @return {Promise}
-     */
-    var setReferencePoint = function (referencePoint) {
-        var dispatchPromise = this.dispatchAsync(events.panel.REFERENCE_POINT_CHANGED, {
-            referencePoint: referencePoint
-        });
-
-        var preferencesPromise = this.transfer(preferences.setPreference,
-            REFERENCE_POINT_PREFS_KEY, referencePoint);
-
-        return Promise.join(dispatchPromise, preferencesPromise);
-    };
-    setReferencePoint.action = {
-        reads: [],
-        writes: [locks.JS_PANEL],
-        transfers: [preferences.setPreference],
-        modal: true
-    };
-
-    /**
-     * Set the UI color stop.
-     *
-     * @param {{stop: string}} payload
-     * @return {Promise}
-     */
-    var setColorStop = function (payload) {
-        var stop = payload.stop,
-            psStop = appLib.colorStops[stop],
-            setColorStop = appLib.setColorStop(psStop),
-            setColorStopPromise = descriptor.playObject(setColorStop),
-            dispatchPromise = this.dispatchAsync(events.panel.COLOR_STOP_CHANGED, {
-                stop: stop
-            });
-
-        return Promise.join(dispatchPromise, setColorStopPromise);
-    };
-    setColorStop.action = {
-        reads: [],
-        writes: [locks.PS_APP, locks.JS_PANEL],
-        transfers: []
-    };
-
-    /**
      * Event handlers initialized in beforeStartup.
      *
      * @private
      * @type {function()}
      */
     var _scrollHandler,
-        _activationChangeHandler,
-        _resizeHandler,
         _displayConfigurationChangedHandler;
 
     /**
      * Register event listeners for UI change events, and initialize the UI.
      *
-     * @param {boolean} reset Indicates whether this is being called as part of a reset
      * @return {Promise}
      */
-    var beforeStartup = function (reset) {
+    var beforeStartup = function () {
         var DEBOUNCE_DELAY = 200;
 
         var setTransformDebounced = synchronization.debounce(function (event) {
@@ -635,22 +368,6 @@ define(function (require, exports) {
         }.bind(this);
         descriptor.addListener("scroll", _scrollHandler);
 
-        // Handles Photoshop focus change events
-        _activationChangeHandler = function (event) {
-            this.dispatchAsync(events.panel.TOGGLE_OVERLAYS, { enabled: event.becameActive });
-        }.bind(this);
-        adapterOS.addListener("activationChanged", _activationChangeHandler);
-
-        var windowResizeDebounced = synchronization.debounce(function () {
-            return this.flux.actions.ui.setOverlayCloaking();
-        }, this, DEBOUNCE_DELAY, false);
-
-        // Handles window resize for resetting superselect tool policies
-        _resizeHandler = function (event) {
-            windowResizeDebounced(event);
-        };
-        window.addEventListener("resize", _resizeHandler);
-
         _displayConfigurationChangedHandler = synchronization.debounce(
             this.flux.actions.ui.handleDisplayConfigurationChanged, this, DEBOUNCE_DELAY);
         adapterOS.addListener("displayConfigurationChanged", _displayConfigurationChangedHandler);
@@ -664,27 +381,12 @@ define(function (require, exports) {
         // Enable target path suppression
         var pathPromise = adapterUI.setSuppressTargetPaths(false);
 
-        // Initialize the reference point from preferences
-        var preferences = this.flux.store("preferences"),
-            referencePoint = preferences.get(REFERENCE_POINT_PREFS_KEY, DEFAULT_REFERENCE_POINT),
-            setReferencePointPromise = this.transfer(setReferencePoint, referencePoint);
-
-        // Initialize the UI color stop
-        const colorStopPromise = uiUtil.getPSColorStop()
-            .bind(this)
-            .then(function (stop) {
-                this.dispatch(events.panel.COLOR_STOP_CHANGED, {
-                    stop: stop
-                });
-            });
-
-        return Promise.join(osPromise, owlPromise, pathPromise, setReferencePointPromise, colorStopPromise)
-            .return(reset);
+        return Promise.join(osPromise, owlPromise, pathPromise);
     };
     beforeStartup.action = {
         reads: [],
-        writes: [locks.PS_APP, locks.JS_PANEL],
-        transfers: [setReferencePoint],
+        writes: [locks.PS_APP],
+        transfers: [],
         modal: true
     };
 
@@ -718,9 +420,7 @@ define(function (require, exports) {
      */
     var onReset = function () {
         descriptor.removeListener("scroll", _scrollHandler);
-        adapterOS.removeListener("activationChanged", _activationChangeHandler);
         adapterOS.removeListener("displayConfigurationChanged", _displayConfigurationChangedHandler);
-        window.removeEventListener("resize", _resizeHandler);
 
         return Promise.resolve();
     };
@@ -730,22 +430,12 @@ define(function (require, exports) {
         modal: true
     };
 
-    exports.enableTooltips = enableTooltips;
-    exports.disableTooltips = disableTooltips;
-    exports.togglePinnedToolbar = togglePinnedToolbar;
-    exports.toggleSingleColumnMode = toggleSingleColumnMode;
     exports.updateTransform = updateTransform;
     exports.setTransform = setTransform;
-    exports.setOverlayCloaking = setOverlayCloaking;
-    exports.cloak = cloak;
-    exports.updatePanelSizes = updatePanelSizes;
-    exports.setOverlayOffsetsForFirstDocument = setOverlayOffsetsForFirstDocument;
     exports.centerBounds = centerBounds;
     exports.centerOn = centerOn;
     exports.zoomInOut = zoomInOut;
     exports.zoom = zoom;
-    exports.setReferencePoint = setReferencePoint;
-    exports.setColorStop = setColorStop;
     exports.handleDisplayConfigurationChanged = handleDisplayConfigurationChanged;
 
     exports.beforeStartup = beforeStartup;
