+++ src/js/jsx/DocumentHeader.jsx
@@ -101,7 +101,7 @@ define(function (require, exports, module) {
                 headerHeight = 0;
             }
 
-            return this.getFlux().actions.ui.updatePanelSizes({
+            return this.getFlux().actions.panel.updatePanelSizes({
                 headerHeight: headerHeight
             });
         },
